requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.2
Flask==3.1.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==25.3.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
python-dotenv==1.2.2
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
